* qt5

+++ trace_server/connection_setup.cpp
@@ -126,7 +126,7 @@ void Connection::handleCSVSetup (QString const & fname)
 	m_table_view_widget->horizontalHeader()->resizeSections(QHeaderView::Fixed);
 
 	columns_sizes_t const & sizes = *sessionState().m_columns_sizes;
-	for (size_t c = 0, ce = sizes.size(); c < ce; ++c)
+	for (int c = 0, ce = sizes.size(); c < ce; ++c)
 	{
 		m_table_view_widget->horizontalHeader()->resizeSection(c, sizes.at(c));
 		//qDebug("sizes: %u %u %u", sizes.at(0), sizes.at(1), sizes.at(2));
